Add PaddleRS at specific commit da1a7457e37b7a70ccf6045f4be08077efa1e0d4

+++ PaddleRS/paddlers/models/ppdet/data/crop_utils/chip_box_utils.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
+# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -128,7 +128,7 @@ def nms(dets, thresh):
     order = scores.argsort()[::-1]
 
     ndets = dets.shape[0]
-    suppressed = np.zeros((ndets), dtype=np.int32)
+    suppressed = np.zeros((ndets), dtype=np.int)
 
     # nominal indices
     # _i, _j
